updated type.c

+++ src/stdlib/type.c
@@ -1,7 +1,7 @@
 // this function returns the type of the given value
 char *type(char *value)
 {
-	int index = strlen(value) - 1;
+	int index = (int) strlen(value) - 1;
 
 	if ((value[0] == '"' && value[index] == '"') ||
 		(value[0] == '\'' && value[index] == '\'') ||
